+++ ts_modules/calendar/frontend_two_stage/logic/meeting_clients/meeting_client_controller.js
@@ -83,17 +83,19 @@ export default class MeetingClientController {
         m.setInfoAlert(`Dieser Termin ist bereits an einen Kunden vergeben. Bearbeiten des Termins ist nur möglich, nachdem die Kundendaten gelöscht wurden.`);
         cv.setAssignDeleteButtonEvent(async () => {
             try {
-                await Alert.deleteConfirm("Kundendaten löschen", "Sollen die Kundendaten wirklich gelöscht werden? Der Kunde und der Berater werden darüber per Mail informiert.");
-                let resp = await CalendarModule.meetingClientService.deleteClient(meetingId);
-                resp.mode = "alert";
-                await Stubegru.utils.wait(200); //Wait until the delete confirm alert is closed
-                Alert.alertResp(resp, "Kundendaten löschen");
-                if (resp.status == "error") {
-                    throw new Error(resp.message);
+                let confirmResp = await Alert.deleteConfirm("Kundendaten löschen", "Sollen die Kundendaten wirklich gelöscht werden? Der Kunde und der Berater werden darüber per Mail informiert.");
+                if (confirmResp.isConfirmed) {
+                    let resp = await CalendarModule.meetingClientService.deleteClient(meetingId);
+                    resp.mode = "alert";
+                    await Stubegru.utils.wait(200); //Wait until the delete confirm alert is closed
+                    Alert.alertResp(resp, "Kundendaten löschen");
+                    if (resp.status == "error") {
+                        throw new Error(resp.message);
+                    }
+                    await CalendarModule.calendarView.refresh();
+                    m.setUnsavedChanges(false);
+                    CalendarModule.meetingController.openFreeMeeting(meetingId);
                 }
-                await CalendarModule.calendarView.refresh();
-                m.setUnsavedChanges(false);
-                CalendarModule.meetingController.openFreeMeeting(meetingId);
             }
             catch (error) {
                 Alert.alertError(error);
