+++ ts_modules/calendar/frontend_two_stage/logic/meetings/meeting_controller.ts
@@ -111,12 +111,14 @@ export default class MeetingController {
 
         m.setFooterDeleteButtonEvent(async () => {
             try {
-                await Alert.deleteConfirm("Termin löschen", "Soll dieser Termin wirklich gelöscht werden?")
-                let resp = await CalendarModule.meetingService.delete(meetingId);
-                Alert.alertResp(resp, "Termin löschen");
-                await CalendarModule.calendarView.refresh();
-                m.setUnsavedChanges(false);
-                m.setModalVisible(false);
+                let confirmResp = await Alert.deleteConfirm("Termin löschen", "Soll dieser Termin wirklich gelöscht werden?");
+                if (confirmResp.isConfirmed) {
+                    let resp = await CalendarModule.meetingService.delete(meetingId);
+                    Alert.alertResp(resp, "Termin löschen");
+                    await CalendarModule.calendarView.refresh();
+                    m.setUnsavedChanges(false);
+                    m.setModalVisible(false);
+                }
             } catch (error) { Alert.alertError(error); }
         });
